% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{add_insert_statement}
\alias{add_insert_statement}
\title{If printed, include an 'INSERT INTO ... VALUES' statement}
\usage{
add_insert_statement(.ddl)
}
\arguments{
\item{.ddl}{An object of class 'ddl'}
}
\description{
If printed, include an 'INSERT INTO ... VALUES' statement
}
