% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{primary_key}
\alias{primary_key}
\alias{not_null}
\alias{default_vals}
\alias{unique_vals}
\alias{check_vals}
\title{Apply constraints to a ddl object from create_table}
\usage{
primary_key(.ddl, v)

not_null(.ddl, v)

default_vals(.ddl, v)

unique_vals(.ddl, v)

check_vals(.ddl, v)
}
\arguments{
\item{.ddl}{An object of type 'ddl'}

\item{v}{A vector or named list}
}
\value{
An object of type 'ddl'
}
\description{
Apply constraints
}
