% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create a new 'ddl' object and statement.}
\usage{
create_table(x, table_name = NULL, if_not_exists = TRUE)
}
\arguments{
\item{x}{A data frame.}

\item{table_name}{Specify the table name. Defaults to object name.}

\item{if_not_exists}{Whether to include an IF NOT EXISTS statement.}
}
\value{
An object of class 'ddl'.
}
\description{
`create_table()` initializes a `ddl` object. `dll` objects create valid DDL statements
 and print it cleanly to the console.
}
\examples{
create_table(mtcars)
}
