% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{dputsql}
\alias{dputsql}
\title{Create table scheme and insert values ddl statements.}
\usage{
dputsql(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
An object of class 'ddl'.
}
\description{
Create table scheme and insert values ddl statements.
}
\examples{
dputsql(mtcars)
}
