% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ddl.R
\name{write_ddl_file}
\alias{write_ddl_file}
\alias{ddl_to_clipboard}
\title{Copy DDL statement to your clipboard or write to a new SQL file}
\usage{
write_ddl_file(.ddl, path, quiet = FALSE)

ddl_to_clipboard(.ddl)
}
\arguments{
\item{.ddl}{An object of type 'ddl'}

\item{path}{Path to target file.}

\item{quiet}{Whether to message about what is happening.}
}
\description{
Copy DDL statement to your clipboard or write to a new SQL file
}
